Move splash screen to its own component
